requirements.txt
# backend stuff
opencv-python-headless==4.12.0.88
numpy==2.2.6
numba==0.63.1
rawpy==0.25.1
imageio==2.37.2
tifffile==2025.12.20
imagecodecs==2026.1.1
Pillow==12.0.0
scipy==1.16.3
matplotlib==3.10.8
Jinja2==3.1.6
# Streamlit ui framework
streamlit==1.52.2
streamlit-image-coordinates==0.4.0
streamlit-javascript==0.1.5
# For tests, type & style checks
pytest==9.0.2
mypy==1.19.1
flake8==7.3.0
ruff==0.14.10
